+++ TVLoader/TVLoaderPlugin.cs
@@ -12,7 +12,7 @@ namespace TVLoader
 	{
 		private const string MyGUID = "rattenbonkers.TVLoader";
 		private const string PluginName = "TVLoader";
-		private const string VersionString = "1.0.3";
+		private const string VersionString = "1.1.0";
 
 		private static readonly Harmony Harmony = new Harmony(MyGUID);
 		public static ManualLogSource Log = new ManualLogSource(PluginName);
@@ -21,10 +21,9 @@ namespace TVLoader
 		{
 			Log = Logger;
 
-
 			Harmony.PatchAll();
 			VideoManager.Load();
-			Logger.LogInfo($"PluginName: {PluginName}, VersionString: {VersionString} is loaded. Video Count: ${VideoManager.Videos.Count}");
+			Logger.LogInfo($"PluginName: {PluginName}, VersionString: {VersionString} is loaded. Video Count: {VideoManager.Videos.Count}");
 		}
 
 	}